Pull redis specific logic out of QueryCounter transform
